Agregados mensajes de confirmación de cerrar y eliminar

+++ trunk/Sigarep.webapp/src/sigarep/viewmodels/maestros/VMTipoMotivo.java
@@ -1,15 +1,24 @@
 package sigarep.viewmodels.maestros;
 import java.util.List;
 
+import org.zkoss.bind.Binder;
+import org.zkoss.bind.annotation.AfterCompose;
 import org.zkoss.bind.annotation.Command;
+import org.zkoss.bind.annotation.ContextParam;
+import org.zkoss.bind.annotation.ContextType;
 import org.zkoss.bind.annotation.Init;
 import org.zkoss.bind.annotation.NotifyChange;
+import org.zkoss.zk.ui.Component;
+import org.zkoss.zk.ui.event.EventListener;
+import org.zkoss.zk.ui.select.Selectors;
 import org.zkoss.zk.ui.select.annotation.VariableResolver;
 import org.zkoss.zk.ui.select.annotation.Wire;
 import org.zkoss.zk.ui.select.annotation.WireVariable;
 
+import org.zkoss.zul.Messagebox;
 import org.zkoss.zul.Textbox;
 import org.zkoss.zul.Window;
+import org.zkoss.zul.Messagebox.ClickEvent;
 
 
 import sigarep.herramientas.MensajesAlUsuario;
@@ -38,6 +47,13 @@ public class VMTipoMotivo {
 	@Wire Textbox txtnombreTipoMotivo;
     @Wire Window winTipoMotivo;
     MensajesAlUsuario mensajeAlUsuario = new MensajesAlUsuario();
+    
+    @Wire("#winTipoMotivo")//para conectarse a la ventana con el ID
+	Window ventana;
+	 @AfterCompose //para poder conectarse con los componentes en la vista, es necesario si no da null Pointer
+    public void afterCompose(@ContextParam(ContextType.VIEW) Component view){
+        Selectors.wireComponents(view, this, false);
+    }
 	
     //Metodos set y get
 
@@ -122,7 +138,7 @@ public class VMTipoMotivo {
     @Command
 	@NotifyChange({"idTipoMotivo","nombreTipoMotivo", "descripcion","estatus","listaTipoMotivo"})//el notifychange le  avisa a que parametros en la pantalla se van a cambiar, en este caso es nombre,apellido,email,sexo se va a colocar en blanco al guardar!!
 	public void guardarTipoMotivo(){
-    	if (nombreTipoMotivo == null || nombreTipoMotivo.equals("") || descripcion.equals("") || descripcion == null) {
+    	if (nombreTipoMotivo == null || descripcion == null) {
 			mensajeAlUsuario.advertenciaLlenarCampos();
 		} else {
 			TipoMotivo tipo = new TipoMotivo(idTipoMotivo, descripcion, true, nombreTipoMotivo, false);
@@ -163,8 +179,8 @@ public class VMTipoMotivo {
 	@NotifyChange({"listaTipoMotivo","idTipoMotivo","nombreTipoMotivo", "estatus","descripcion","nombreFiltro"})
 	public void limpiar(){
     	idTipoMotivo= null;
-    	nombreTipoMotivo = "";
-		descripcion="";
+    	nombreTipoMotivo = null;
+		descripcion=null;
 		nombreFiltro= "";
 		listaTipoMotivo();
 		
@@ -180,18 +196,34 @@ public class VMTipoMotivo {
 	 * @throws Debe
 	 *             seleccionar un registro para poder eliminarlo
 	 */
-  	@Command
-  	@NotifyChange({"listaTipoMotivo", "idTipoMotivo","nombreTipoMotivo", "descripcion"})
-  	public void eliminarTipoMotivo(){
-  		if (nombreTipoMotivo==null || nombreTipoMotivo.equals("") || descripcion==null || descripcion.equals("") ){
-  			mensajeAlUsuario.advertenciaSeleccionarParaEliminar();
+  	@SuppressWarnings("unchecked")
+	@Command
+	@NotifyChange({"listaTipoMotivo", "idTipoMotivo","nombreTipoMotivo", "descripcion"})
+	public void eliminarTipoMotivo(@ContextParam(ContextType.BINDER) final Binder binder){
+  		if (nombreTipoMotivo==null  || descripcion==null ) {
+			mensajeAlUsuario.advertenciaSeleccionarParaEliminar();
+		} else {
+			Messagebox.show("¿Desea eliminar el registro realmente?","Confirmar",new Messagebox.Button[] { Messagebox.Button.YES,Messagebox.Button.NO },
+					Messagebox.QUESTION,new EventListener<ClickEvent>() {
+				@SuppressWarnings("incomplete-switch")
+				public void onEvent(ClickEvent e) throws Exception {
+					switch (e.getButton()) {
+						case YES:
+							//if you call super.delete here, since original zk event is not control by binder
+							//the change of viewmodel will not update to the ui.
+							//so, I post a delete to trigger to process it in binder controll.
+							//binder.postCommand("limpiar", null);
+							serviciotipomotivo.eliminarTipoMotivo(getTiposeleccionado().getIdTipoMotivo());
+							mensajeAlUsuario.informacionEliminarCorrecto();
+							binder.postCommand("limpiar", null);
+						case NO:
+					
+							binder.postCommand("limpiar", null);
+					}
+				}
+			});		
 		}
-		else{
-  		serviciotipomotivo.eliminarTipoMotivo(getTiposeleccionado().getIdTipoMotivo());
-  		mensajeAlUsuario.informacionEliminarCorrecto();
-  		limpiar();
-  	}
-  	}
+	}
   	
   	
   	/**
@@ -226,6 +258,50 @@ public class VMTipoMotivo {
 	@NotifyChange({ "listaTipoMotivo", "nombreFiltro" })
 	public void filtros() {
 		listaTipoMotivo = serviciotipomotivo.buscarTipoMotivo(nombreFiltro);
+	}
+	
+	/**
+	 * Cerrar Ventana
+	 * 
+	 * @param binder
+	 * @return cierra el .zul asociado al VM
+	 * @throws No
+	 *             dispara ninguna excepcion.
+	 */
+	
+	@SuppressWarnings("unchecked")
+	@Command
+	@NotifyChange({"listaTipoMotivo", "idTipoMotivo","nombreTipoMotivo", "descripcion"})
+	public void cerrarVentana(@ContextParam(ContextType.BINDER) final Binder binder){
+			
+		if (nombreTipoMotivo !=null  || descripcion !=null ) 
+		{
+			Messagebox.show("¿Realemente desea cerrar la ventana sin guardar los cambios?","Confirmar",new Messagebox.Button[] { Messagebox.Button.YES,Messagebox.Button.NO },
+					Messagebox.QUESTION,new EventListener<ClickEvent>() {
+				@SuppressWarnings("incomplete-switch")
+				public void onEvent(ClickEvent e) throws Exception {
+					switch (e.getButton()) {
+						case YES:
+								ventana.detach();
+					
+					}
+				}
+			});		
+		}
+		else{
+		Messagebox.show("¿Realmente desea cerrar la ventana?","Confirmar",new Messagebox.Button[] { Messagebox.Button.YES,Messagebox.Button.NO },
+					Messagebox.QUESTION,new EventListener<ClickEvent>() {
+				@SuppressWarnings("incomplete-switch")
+				public void onEvent(ClickEvent e) throws Exception {
+					switch (e.getButton()) {
+						case YES:
+								ventana.detach();
+					
+					
+					}
+				}
+			});		
+		}
 	}
