+++ trunk/Sigarep.webapp/src/sigarep/viewmodels/maestros/VMpreguntaBasica.java
@@ -1,10 +1,23 @@
 package sigarep.viewmodels.maestros;
 import java.util.List;
+
+import org.zkoss.bind.Binder;
+import org.zkoss.bind.annotation.AfterCompose;
 import org.zkoss.bind.annotation.Command;
+import org.zkoss.bind.annotation.ContextParam;
+import org.zkoss.bind.annotation.ContextType;
 import org.zkoss.bind.annotation.Init;
 import org.zkoss.bind.annotation.NotifyChange;
+import org.zkoss.zk.ui.Component;
+import org.zkoss.zk.ui.event.EventListener;
+import org.zkoss.zk.ui.select.Selectors;
 import org.zkoss.zk.ui.select.annotation.VariableResolver;
+import org.zkoss.zk.ui.select.annotation.Wire;
 import org.zkoss.zk.ui.select.annotation.WireVariable;
+import org.zkoss.zul.Messagebox;
+import org.zkoss.zul.Window;
+import org.zkoss.zul.Messagebox.ClickEvent;
+
 import sigarep.herramientas.MensajesAlUsuario;
 import sigarep.modelos.data.maestros.PreguntaBasica;
 import sigarep.modelos.servicio.maestros.ServicioPreguntaBasica;
@@ -25,6 +38,13 @@ public class VMpreguntaBasica {
 	private PreguntaBasica preguntaseleccionada;
 	MensajesAlUsuario mensajeAlUsuario = new MensajesAlUsuario (); // Instancia de la Clase de mensajes 
     
+	@Wire("#winActualizarPreguntasBasicas")//para conectarse a la ventana con el ID
+	Window ventana;
+	 @AfterCompose //para poder conectarse con los componentes en la vista, es necesario si no da null Pointer
+    public void afterCompose(@ContextParam(ContextType.VIEW) Component view){
+        Selectors.wireComponents(view, this, false);
+    }
+	 
 	//Metodos Get y Set de la clase 
     public Integer getIdPreguntaBasica() {
 		return idPreguntaBasica;
@@ -85,7 +105,7 @@ public class VMpreguntaBasica {
 	@Command // Permite manipular la propiedad de ViewModel
 	@NotifyChange({"id_pregunta_basica", "pregunta", "respuesta","estatus","listaPregunta"})//el notifychange le  avisa a que parametros en la pantalla se van a cambiar, en este caso es los atributos de la pantalla se va a colocar en blanco al guardar!!
 	public void guardarPregunta(){
-		if (pregunta.equals("")||respuesta.equals(""))
+		if (pregunta==null||respuesta==null)
 			mensajeAlUsuario.advertenciaLlenarCampos();
 		else{
 		PreguntaBasica preb = new PreguntaBasica (idPreguntaBasica,pregunta,respuesta,true);
@@ -99,9 +119,11 @@ public class VMpreguntaBasica {
 	 * @return No devuelve ningun valor.
 	 */
 	@Command
-	@NotifyChange({"id_pregunta_basica", "pregunta", "respuesta","estatus"})
+	@NotifyChange({"listaPregunta", "idPreguntaBasica ", "pregunta", "respuesta","estatus"})
 	public void limpiar(){
-		 pregunta = "";respuesta="";
+		idPreguntaBasica = null;
+		pregunta = null; 
+		 respuesta =null;
 		 buscarPregunta ();
 	}
 
@@ -118,17 +140,34 @@ public class VMpreguntaBasica {
 	 * @return No devuelve ningun valor.
 	 * @throws la Excepcion es que quiera eliminar con los campos vacios, sin seleccionar ningun registro
 	 */
+	
+	@SuppressWarnings("unchecked")
 	@Command
 	@NotifyChange({"listaPregunta","pregunta","respuesta"})
-	public void eliminarPreguntaBasica(){
-		if (pregunta.equals("")|| respuesta.equals("")){
+	public void  eliminarPreguntaBasica(@ContextParam(ContextType.BINDER) final Binder binder){
+		if (pregunta == null|| respuesta == null){
 			mensajeAlUsuario.advertenciaSeleccionarParaEliminar();
-  		}
-		else{
-		serviciopreguntabasica.eliminarPregunta(getPreguntaseleccionada().getIdPreguntaBasica());
-		limpiar();
-		mensajeAlUsuario.informacionEliminarCorrecto();
-	}
+		} else {
+			Messagebox.show("¿Desea eliminar el registro realmente?","Confirmar",new Messagebox.Button[] { Messagebox.Button.YES,Messagebox.Button.NO },
+					Messagebox.QUESTION,new EventListener<ClickEvent>() {
+				@SuppressWarnings("incomplete-switch")
+				public void onEvent(ClickEvent e) throws Exception {
+					switch (e.getButton()) {
+						case YES:
+							//if you call super.delete here, since original zk event is not control by binder
+							//the change of viewmodel will not update to the ui.
+							//so, I post a delete to trigger to process it in binder controll.
+							//binder.postCommand("limpiar", null);
+							serviciopreguntabasica.eliminarPregunta(getPreguntaseleccionada().getIdPreguntaBasica());
+							mensajeAlUsuario.informacionEliminarCorrecto();
+							binder.postCommand("limpiar", null);
+						case NO:
+					
+							binder.postCommand("limpiar", null);
+					}
+				}
+			});		
+		}
 	}
 	/** mostrarSeleccionada
 	 * @param pregunta, respuesta.
@@ -141,4 +180,50 @@ public class VMpreguntaBasica {
 		pregunta=getPreguntaseleccionada().getPregunta();
 		respuesta=getPreguntaseleccionada().getRespuesta();
 	}
+	
+	/**
+	 * Cerrar Ventana
+	 * 
+	 * @param binder
+	 * @return cierra el .zul asociado al VM
+	 * @throws No
+	 *             dispara ninguna excepcion.
+	 */
+	@SuppressWarnings("unchecked")
+	@Command
+	@NotifyChange({"listaPregunta","pregunta","respuesta"})
+	public void cerrarVentana(@ContextParam(ContextType.BINDER) final Binder binder){
+			
+		if (pregunta != null|| respuesta != null) 
+		{
+			Messagebox.show("¿Realmente desea cerrar la ventana sin guardar los cambios?","Confirmar",new Messagebox.Button[] { Messagebox.Button.YES,Messagebox.Button.NO },
+					Messagebox.QUESTION,new EventListener<ClickEvent>() {
+				@SuppressWarnings("incomplete-switch")
+				public void onEvent(ClickEvent e) throws Exception {
+					switch (e.getButton()) {
+						case YES:
+								ventana.detach();
+					
+					}
+				}
+			});		
+		}
+		else{
+		Messagebox.show("¿Realmente desea cerrar la ventana?","Confirmar",new Messagebox.Button[] { Messagebox.Button.YES,Messagebox.Button.NO },
+					Messagebox.QUESTION,new EventListener<ClickEvent>() {
+				@SuppressWarnings("incomplete-switch")
+				public void onEvent(ClickEvent e) throws Exception {
+					switch (e.getButton()) {
+						case YES:
+								ventana.detach();
+					
+					
+					}
+				}
+			});		
+		}
+	}
+	
+	
+	
 }
